python-proxy/requirements.txt
fastapi==0.105.0
uvicorn==0.24.0
python-dotenv==1.0.0
google-generativeai==0.3.1
httpx==0.25.1
pydantic==2.5.2
python-multipart==0.0.6
aiofiles==23.2.1
jinja2==3.1.2
websockets==12.0
starlette==0.27.0
